[*] Создал структуру папок страницы Game добавил для неё роуты
