+++ src/routes/Game/routes/Finish/index.js
@@ -1,0 +1,11 @@
+import s from './style.module.css';
+
+const FinishPage = () => {
+    return (
+        <div className={s.root}>
+          <p>Game over</p>
+        </div>
+    );
+};
+
+export default FinishPage;
